+++ CHANGELOG.md
@@ -85,6 +85,11 @@
 - add `src/MessageView.svelte`
 - `/src/icons/Icon.svelte`: add the `glyphSize` prop so that we can pass glyphs with sizes different from 24x24
 - add `src/icons/svizzle/A11yPerson.svelte`
+- `ScreenGauge`:
+	- add `landscape`|`portrait` to `classes`
+	- changed the output shape:
+		- remame `orientationFlags` to `orientations`
+		- remame `sizeFlags` to `sizes`
 - updated some dev dependencies
 
 ## `@svizzle/utils` v0.14.0 (next)